+++ test/integration/suite/02-language-features.test.ts
@@ -317,7 +317,7 @@ suite("Language feature integration", () => {
       });
 
       test("diagnostics resolve split haproxy.d layouts with configured globs", async () => {
-        await updateHaproxySetting("workspaceSymbols.maxFiles", 300);
+        await updateHaproxySetting("workspaceSymbols.maxFiles", 1000);
         await updateHaproxySetting("workspaceSymbols.maxTotalLines", 100000);
         await updateHaproxySetting("workspaceSymbols.include", [
           "**/haproxy-tests/haproxy.d/**/*.cfg",